DMP 1890 Download Annotation Document
Moved data creation to DartsDatabaseStub
